[Improvement-3369][api] Introduce session service interface for clear code (#3408)
* [Improvement][api] Introduce session service interface for clear code

* Update SessionService.java

* [Improvement][api] Add @Service

* Add licenses
